Physics system, updates to scene graph system, music system, interpolation animators

+++ app/src/main/cpp/core/Vector4.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <cmath>
-#include <math.h>
+#include "Math2D.h"
 
 namespace core
 {
@@ -111,15 +111,15 @@ namespace core
 
 		inline float length() const
 		{
-			return std::sqrtf(x * x + y * y + z * z + w * w);
+			return std::sqrt(x * x + y * y + z * z + w * w);
 		};
 
 		inline void normalise()
 		{
 			float lenSq = x * x + y * y + z * z + w * w;
-			if (fabsf(lenSq - 1.0f) > EPSILON)
+			if (std::fabs(lenSq - 1.0f) > EPSILON)
 			{
-				*this /= std::sqrtf(lenSq);
+				*this /= std::sqrt(lenSq);
 			}
 
 		};
